Cập nhật code mới

+++ webtraodoidodung/app1/models.py
@@ -21,7 +21,7 @@ class CreateUserForm(UserCreationForm):
 class Product(models.Model):
     category = models.ManyToManyField(Category,related_name='product')
     name = models.CharField(max_length=200,null=True)
-    price = models.FloatField()
+    price = models.PositiveIntegerField()
     digital = models.BooleanField(default=False,null=True,blank=False)
     image = models.ImageField(null=True,blank=True)
     detail = models.TextField(null=True,blank=True)
